NetworkPkg: Fix loop index type mismatches
Use loop index types that can represent the corresponding count or
buffer size. This prevents truncation and wraparound when loop bounds
exceed the range of the narrower index type.

+++ NetworkPkg/WifiConnectionManagerDxe/WifiConnectionMgrHiiConfigAccess.c
@@ -268,7 +268,7 @@ WifiMgrGetStrAKMList (
   IN  WIFI_MGR_NETWORK_PROFILE  *Profile
   )
 {
-  UINT8   Index;
+  UINT16  Index;
   UINT16  AKMSuiteCount;
   CHAR16  *AKMListDisplay;
   UINTN   Length;
@@ -327,7 +327,7 @@ WifiMgrGetStrCipherList (
   IN  WIFI_MGR_NETWORK_PROFILE  *Profile
   )
 {
-  UINT8   Index;
+  UINT16  Index;
   UINT16  CipherSuiteCount;
   CHAR16  *CipherListDisplay;
